initial clippy tests
